AK: Remove unnecessary casts to size_t, after Vector changes
Now that Vector uses size_t, we can remove a whole bunch of redundant
casts to size_t.

+++ AK/String.cpp
@@ -143,7 +143,7 @@ Vector<String> String::split_limit(char separator, size_t limit, bool keep_empty
 
     Vector<String> v;
     size_t substart = 0;
-    for (size_t i = 0; i < length() && ((size_t)v.size() + 1) != limit; ++i) {
+    for (size_t i = 0; i < length() && (v.size() + 1) != limit; ++i) {
         char ch = characters()[i];
         if (ch == separator) {
             size_t sublen = i - substart;
